+++ NewsParser.py
@@ -2,15 +2,13 @@
 # -*- coding: utf-8 -*-
 # @Author: balicanta
 # @Date:   2014-10-25 00:09:39
-# @Last Modified by:   bustta
-# @Last Modified time: 2014-10-27 23:03:54
+# @Last Modified by:   balicanta
+# @Last Modified time: 2014-11-01 21:09:55
 
 import sys
 
-from UdnNewsParseStrategy import UdnNewsParseStrategy
-from WorldYamParseStrategy import WorldYamParseStrategy
-from LtnNewsParseStrategy import LtnNewsParseStrategy
-from AbstractNewsParseStrategy import AbstractNewsParseStrategy
+from strategies.LtnNewsParseStrategy import LtnNewsParseStrategy
+from strategies.AbstractNewsParseStrategy import AbstractNewsParseStrategy
 
 from requests.utils import get_encodings_from_content
 from bs4 import BeautifulSoup
@@ -24,7 +22,6 @@ class NewsParser():
     parse_strategy = None
 
     def __init__(self, URL):
-        from AbstractNewsParseStrategy import AbstractNewsParseStrategy
         self.url = URL
         for parse_strategy in vars()['AbstractNewsParseStrategy'].__subclasses__():
             if parse_strategy().isURLMatch(URL):
