+++ __init__.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# @Author: balicanta
+# @Date:   2014-11-01 21:02:00
+# @Last Modified by:   balicanta
+# @Last Modified time: 2014-11-01 21:02:00
