+++ strategies/AbstractNewsParseStrategy.py
