+++ strategies/WorldYamParseStrategy.py
